Swamy/remove unused using and refactoring entity classes (#1371)
* Removed Unused Usings

* Code Refactoring for readbility. Added spaces and moved the classes to separate file
