+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Config/UrlsConfig.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Config
 {
@@ -10,12 +7,14 @@ namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Config
         public class CatalogOperations
         {
             public static string GetItemById(int id) => $"/api/v1/catalog/items/{id}";
+
             public static string GetItemsById(IEnumerable<int> ids) => $"/api/v1/catalog/items?ids={string.Join(',', ids)}";
         }
 
         public class BasketOperations
         {
             public static string GetItemById(string id) => $"/api/v1/basket/{id}";
+
             public static string UpdateBasket() => "/api/v1/basket";
         }
 
@@ -25,10 +24,15 @@ namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Config
         }
 
         public string Basket { get; set; }
+
         public string Catalog { get; set; }
+
         public string Orders { get; set; }
+
         public string GrpcBasket { get; set; }
+
         public string GrpcCatalog { get; set; }
+
         public string GrpcOrdering { get; set; }
     }
 }
