+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Models/AddBasketItemRequest.cs
@@ -1,13 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
+﻿namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
 {
     public class AddBasketItemRequest
     {
         public int CatalogItemId { get; set; }
+
         public string BasketId { get; set; }
 
         public int Quantity { get; set; }
