+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Models/BasketData.cs
@@ -2,6 +2,7 @@
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
 {
+
     public class BasketData
     {
         public string BuyerId { get; set; }
@@ -10,7 +11,6 @@ namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
 
         public BasketData()
         {
-
         }
 
         public BasketData(string buyerId)
@@ -19,15 +19,4 @@ namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
         }
     }
 
-    public class BasketDataItem
-    {
-        public string Id { get; set; }
-        public int ProductId { get; set; }
-        public string ProductName { get; set; }
-        public decimal UnitPrice { get; set; }
-        public decimal OldUnitPrice { get; set; }
-        public int Quantity { get; set; }
-        public string PictureUrl { get; set; }
-
-    }
 }
