+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Models/OrderData.cs
@@ -3,23 +3,39 @@ using System.Collections.Generic;
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
 {
+
     public class OrderData
     {
         public string OrderNumber { get; set; }
+
         public DateTime Date { get; set; }
+
         public string Status { get; set; }
+
         public decimal Total { get; set; }
+
         public string Description { get; set; }
+
         public string City { get; set; }
+
         public string Street { get; set; }
+
         public string State { get; set; }
+
         public string Country { get; set; }
+
         public string ZipCode { get; set; }
+
         public string CardNumber { get; set; }
+
         public string CardHolderName { get; set; }
+
         public bool IsDraft { get; set; }
+
         public DateTime CardExpiration { get; set; }
+
         public string CardExpirationShort { get; set; }
+
         public string CardSecurityNumber { get; set; }
 
         public int CardTypeId { get; set; }
@@ -28,4 +44,5 @@ namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
 
         public List<OrderItemData> OrderItems { get;  } = new List<OrderItemData>();
     }
+
 }
