+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Models/UpdateBasketItemData.cs
@@ -1,0 +1,16 @@
+﻿namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
+{
+
+    public class UpdateBasketItemData
+    {
+        public string BasketItemId { get; set; }
+
+        public int NewQty { get; set; }
+
+        public UpdateBasketItemData()
+        {
+            NewQty = 0;
+        }
+    }
+
+}
