+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Models/UpdateBasketItemsRequest.cs
@@ -2,6 +2,7 @@
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
 {
+
     public class UpdateBasketItemsRequest
     {
         
@@ -15,14 +16,4 @@ namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
         }
     }
 
-    public class UpdateBasketItemData
-    {
-        public string BasketItemId { get; set; }
-        public int NewQty { get; set; }
-
-        public UpdateBasketItemData()
-        {
-            NewQty = 0;
-        }
-    }
 }
