+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Models/UpdateBasketRequest.cs
@@ -2,6 +2,7 @@
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
 {
+    
     public class UpdateBasketRequest
     {
         public string BuyerId { get; set; }
@@ -9,10 +10,4 @@ namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models
         public IEnumerable<UpdateBasketRequestItemData> Items { get; set; }
     }
 
-    public class UpdateBasketRequestItemData
-    {
-        public string Id { get; set; }          // Basket id
-        public int ProductId { get; set; }      // Catalog item id
-        public int Quantity { get; set; }       // Quantity
-    }
 }
