+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Program.cs
@@ -1,10 +1,8 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.Configuration;
 using Serilog;
 using System.IO;
-using System.Net;
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator
 {
