+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Services/BasketService.cs
@@ -1,14 +1,12 @@
 ﻿using Grpc.Net.Client;
+using GrpcBasket;
 using Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Config;
 using Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using System;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
-using GrpcBasket;
-using Grpc.Core;
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Services
 {
