+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Services/GrpcCallerService.cs
@@ -1,9 +1,8 @@
-﻿using System.Net.Http;
-using System.Threading.Tasks;
-using System;
-using Grpc.Core;
-using Serilog;
+﻿using Grpc.Core;
 using Grpc.Net.Client;
+using Serilog;
+using System;
+using System.Threading.Tasks;
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Services
 {
