+++ src/ApiGateways/Mobile.Bff.Shopping/aggregator/Services/OrderingService.cs
@@ -1,4 +1,4 @@
-﻿using Grpc.Net.Client;
+﻿using GrpcOrdering;
 using Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Config;
 using Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Models;
 using Microsoft.Extensions.Logging;
@@ -6,7 +6,6 @@ using Microsoft.Extensions.Options;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
-using GrpcOrdering;
 
 namespace Microsoft.eShopOnContainers.Mobile.Shopping.HttpAggregator.Services
 {
