+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Filters/AuthorizeCheckOperationFilter.cs
@@ -1,11 +1,11 @@
-﻿namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Filters
-{
-    using Microsoft.AspNetCore.Authorization;
-    using Microsoft.OpenApi.Models;
-    using Swashbuckle.AspNetCore.SwaggerGen;
-    using System.Collections.Generic;
-    using System.Linq;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.OpenApi.Models;
+using Swashbuckle.AspNetCore.SwaggerGen;
+using System.Collections.Generic;
+using System.Linq;
 
+namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Filters
+{
     namespace Basket.API.Infrastructure.Filters
     {
         public class AuthorizeCheckOperationFilter : IOperationFilter
