+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/BasketDataItem.cs
@@ -1,17 +1,19 @@
 ﻿namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
 {
 
-    public class OrderItemData
+    public class BasketDataItem
     {
+        public string Id { get; set; }
+
         public int ProductId { get; set; }
 
         public string ProductName { get; set; }
 
         public decimal UnitPrice { get; set; }
 
-        public decimal Discount { get; set; }
+        public decimal OldUnitPrice { get; set; }
 
-        public int Units { get; set; }
+        public int Quantity { get; set; }
 
         public string PictureUrl { get; set; }
     }
