+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/CatalogItem.cs
@@ -1,5 +1,6 @@
 ﻿namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
 {
+
     public class CatalogItem
     {
         public int Id { get; set; }
@@ -10,4 +11,5 @@
 
         public string PictureUri { get; set; }
     }
+
 }
