+++ src/ApiGateways/Web.Bff.Shopping/aggregator/Models/UpdateBasketItemsRequest.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
 {
+
     public class UpdateBasketItemsRequest
     {
-        
         public string BasketId { get; set; }
 
         public ICollection<UpdateBasketItemData> Updates { get; set; }
@@ -18,14 +15,4 @@ namespace Microsoft.eShopOnContainers.Web.Shopping.HttpAggregator.Models
         }
     }
 
-    public class UpdateBasketItemData
-    {
-        public string BasketItemId { get; set; }
-        public int NewQty { get; set; }
-
-        public UpdateBasketItemData()
-        {
-            NewQty = 0;
-        }
-    }
 }
